toolkit/mozapps/downloads/tests/chrome/test_close_download_manager.xul

<?xml version="1.0"?>
<!-- This Source Code Form is subject to the terms of the Mozilla Public
   - License, v. 2.0. If a copy of the MPL was not distributed with this
   - file, You can obtain one at http://mozilla.org/MPL/2.0/.  -->
<!--
 * This test basically checks if the download manager
 * closes when you press the esc key and accel + w.
-->

<window title="Download Manager Test"
        onload="runTest();"
        xmlns:html="http://www.w3.org/1999/xhtml"
        xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul">

  <script type="application/javascript"
          src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"/>
  <script type="application/javascript"
          src="chrome://mochikit/content/tests/SimpleTest/EventUtils.js"/>
  <script type="application/javascript"
          src="utils.js"/>

  <script type="application/javascript">

  <![CDATA[

const dmui = getDMUI();

function testCloseDMWithEscKey(aWin)
{
  function dmWindowClosedListener() {
    aWin.removeEventListener("unload", dmWindowClosedListener, false);
    ok(!dmui.visible, "DMUI closes with ESC key");
    dmui.show();
  }
  aWin.addEventListener("unload", dmWindowClosedListener, false);

  synthesizeKey("VK_ESCAPE", { }, aWin);
}

function testCloseDMWithAccelKey(aWin)
{
  function dmWindowClosedListener() {
    aWin.removeEventListener("unload", dmWindowClosedListener, false);
    ok(!dmui.visible, "DMUI closes with accel + w");
    setCleanState();
    SimpleTest.finish();
  }
  aWin.addEventListener("unload", dmWindowClosedListener, false);

  synthesizeKey("w", { accelKey: true }, aWin);
}

function runTest()
{
  if (!dmui) {
    todo(false, "skip test for toolkit download manager UI");
    return;
  }

  const DLMGR_UI_DONE = "download-manager-ui-done";

  // Close the DM UI if it is already open
  let dm_win = getDMWindow();
  if (dm_win) dm_win.close();

  let testPhase = 0;
  // Specify an observer that will be notified when the dm has been rendered on screen
  let obs = Cc["@mozilla.org/observer-service;1"].
            getService(Ci.nsIObserverService);
  let testObs = {
    observe: function(aSubject, aTopic, aData) {
      SimpleTest.waitForFocus(function () { closeDM(aSubject) }, aSubject);
    }
  };

  function closeDM(win) {
    switch(testPhase++) {
      case 0:
        testCloseDMWithEscKey(win);
        break;
      case 1:
        obs.removeObserver(testObs, DLMGR_UI_DONE);
        testCloseDMWithAccelKey(win);
    }
  }
  obs.addObserver(testObs, DLMGR_UI_DONE, false);

  dmui.show();

  SimpleTest.waitForExplicitFinish();
}

  ]]>
  </script>

  <body xmlns="http://www.w3.org/1999/xhtml">
    <p id="display"></p>
    <div id="content" style="display: none"></div>
    <pre id="test"></pre>
  </body>
</window>
